Remove `App` and put boilerplate back in example
