+++ examples/render_inp.rs
@@ -1,19 +1,36 @@
 use std::{
+    env,
     error::Error,
+    ffi::CString,
     fs::File,
     io::{BufReader, Read},
+    num::NonZeroU32,
 };
 
-use glutin::surface::GlSurface;
-use inox2d::{formats::inp::parse_inp, renderers::opengl::opengl_app, texture::decode_textures};
+use glow::HasContext;
 
-use tracing::{debug, info};
+use glutin::{
+    context::{ContextApi, ContextAttributesBuilder, PossiblyCurrentContext, Version},
+    display::Display,
+    display::GetGlDisplay,
+    prelude::{GlConfig, GlDisplay, NotCurrentGlContextSurfaceAccessor},
+    surface::{GlSurface, Surface, SurfaceAttributesBuilder, WindowSurface},
+};
+
+use glutin_winit::ApiPrefence;
+use inox2d::{
+    formats::inp::parse_inp, renderers::opengl::opengl_renderer, texture::decode_textures,
+};
+use raw_window_handle::HasRawWindowHandle;
+
+use tracing::{debug, error, info, warn};
 
 use tracing_subscriber::{filter::LevelFilter, fmt, prelude::*};
 
 use winit::{
-    event::{Event, StartCause},
-    window::WindowBuilder,
+    event::{ElementState, Event, KeyboardInput, StartCause, VirtualKeyCode, WindowEvent},
+    event_loop::{ControlFlow, EventLoop},
+    window::Window,
 };
 
 use std::path::PathBuf;
@@ -54,25 +71,26 @@ fn main() -> Result<(), Box<dyn Error>> {
     let rx = decode_textures(&mut model.textures);
 
     info!("Setting up windowing and OpenGL");
-    let events = winit::event_loop::EventLoop::new();
+    let App {
+        gl,
+        gl_ctx,
+        gl_surface,
+        gl_display,
+        events,
+        window,
+    } = launch_opengl_window()?;
 
-    let window = WindowBuilder::new()
-        .with_transparent(true)
-        .with_resizable(false)
-        .with_inner_size(winit::dpi::PhysicalSize::new(2048, 2048))
-        .with_title("Render Inochi2D Puppet")
-        .build(&events)
-        .unwrap();
+    info!("Initializing Inox2D renderer");
+    let mut renderer = opengl_renderer(gl, puppet.nodes);
+    renderer.upload_textures(rx, n_textures);
+    info!("Inox2D renderer initialized");
 
-    let mut app = opengl_app(&window, puppet.nodes).unwrap();
-    app.renderer.upload_textures(rx, n_textures);
-
-    let zsorted_nodes = app.renderer.nodes.zsorted();
+    let zsorted_nodes = renderer.nodes.zsorted();
 
     // Event loop
     events.run(move |event, _, control_flow| {
         // They need to be present
-        let _gl_display = &app.display;
+        let _gl_display = &gl_display;
         let _window = &window;
 
         control_flow.set_wait();
@@ -81,13 +99,154 @@ fn main() -> Result<(), Box<dyn Error>> {
             Event::NewEvents(StartCause::Poll) | Event::RedrawRequested(_) => {
                 debug!("Redrawing");
 
-                app.renderer.clear();
-                app.renderer.render_nodes(&zsorted_nodes);
+                renderer.clear();
+                renderer.render_nodes(&zsorted_nodes);
 
-                app.surface.swap_buffers(&app.gl_ctx).unwrap();
+                gl_surface.swap_buffers(&gl_ctx).unwrap();
                 // window.request_redraw();
             }
-            _ => app.update(event, control_flow),
+            _ => handle_event(event, control_flow, &renderer.gl, &gl_surface, &gl_ctx),
         }
     })
+}
+
+struct App {
+    pub gl: glow::Context,
+    pub gl_ctx: PossiblyCurrentContext,
+    pub gl_surface: Surface<WindowSurface>,
+    pub gl_display: Display,
+    pub window: Window,
+    pub events: EventLoop<()>,
+}
+
+fn launch_opengl_window() -> Result<App, Box<dyn Error>> {
+    if cfg!(target_os = "linux") {
+        // disables vsync sometimes on x11
+        if env::var("vblank_mode").is_err() {
+            env::set_var("vblank_mode", "0");
+        }
+    }
+
+    let events = winit::event_loop::EventLoop::new();
+
+    let window_builder = winit::window::WindowBuilder::new()
+        .with_transparent(true)
+        .with_resizable(false)
+        .with_inner_size(winit::dpi::PhysicalSize::new(2048, 2048))
+        .with_title("Render Inochi2D Puppet");
+
+    let (window, gl_config) = glutin_winit::DisplayBuilder::new()
+        .with_preference(ApiPrefence::FallbackEgl)
+        .with_window_builder(Some(window_builder))
+        .build(&events, <_>::default(), |configs| {
+            configs
+                .filter(|c| c.srgb_capable())
+                .max_by_key(|c| c.num_samples())
+                .unwrap()
+        })?;
+
+    let window = window.unwrap(); // set in display builder
+    let raw_window_handle = window.raw_window_handle();
+    let gl_display = gl_config.display();
+
+    let context_attributes = ContextAttributesBuilder::new()
+        .with_context_api(ContextApi::OpenGl(Some(Version::new(3, 1))))
+        .with_profile(glutin::context::GlProfile::Core)
+        .build(Some(raw_window_handle));
+
+    let dimensions = window.inner_size();
+
+    let (gl_surface, gl_ctx) = {
+        let attrs = SurfaceAttributesBuilder::<glutin::surface::WindowSurface>::new().build(
+            raw_window_handle,
+            NonZeroU32::new(dimensions.width).unwrap(),
+            NonZeroU32::new(dimensions.height).unwrap(),
+        );
+
+        let surface = unsafe { gl_display.create_window_surface(&gl_config, &attrs)? };
+        let context = unsafe { gl_display.create_context(&gl_config, &context_attributes)? }
+            .make_current(&surface)?;
+        (surface, context)
+    };
+
+    // Load the OpenGL function pointers
+    let gl = unsafe {
+        glow::Context::from_loader_function(|symbol| {
+            gl_display.get_proc_address(&CString::new(symbol).unwrap()) as *const _
+        })
+    };
+
+    unsafe {
+        gl.debug_message_callback(|_src, ty, _id, sevr, msg| {
+            let ty = match ty {
+                glow::DEBUG_TYPE_ERROR => "Error: ",
+                glow::DEBUG_TYPE_DEPRECATED_BEHAVIOR => "Deprecated Behavior: ",
+                glow::DEBUG_TYPE_MARKER => "Marker: ",
+                glow::DEBUG_TYPE_OTHER => "",
+                glow::DEBUG_TYPE_POP_GROUP => "Pop Group: ",
+                glow::DEBUG_TYPE_PORTABILITY => "Portability: ",
+                glow::DEBUG_TYPE_PUSH_GROUP => "Push Group: ",
+                glow::DEBUG_TYPE_UNDEFINED_BEHAVIOR => "Undefined Behavior: ",
+                glow::DEBUG_TYPE_PERFORMANCE => "Performance: ",
+                ty => unreachable!("unknown debug type {ty}"),
+            };
+            match sevr {
+                glow::DEBUG_SEVERITY_NOTIFICATION => debug!(target: "opengl", "{ty}{msg}"),
+                glow::DEBUG_SEVERITY_LOW => info!(target: "opengl", "{ty}{msg}"),
+                glow::DEBUG_SEVERITY_MEDIUM => warn!(target: "opengl", "{ty}{msg}"),
+                glow::DEBUG_SEVERITY_HIGH => error!(target: "opengl", "{ty}{msg}"),
+                sevr => unreachable!("unknown debug severity {sevr}"),
+            };
+        });
+
+        gl.enable(glow::DEBUG_OUTPUT);
+    }
+
+    unsafe { gl.viewport(0, 0, 2048, 2048) };
+
+    Ok(App {
+        gl,
+        gl_ctx,
+        gl_surface,
+        gl_display,
+        window,
+        events,
+    })
+}
+
+fn handle_event(
+    event: Event<()>,
+    control_flow: &mut ControlFlow,
+    _gl: &glow::Context,
+    gl_surface: &Surface<WindowSurface>,
+    gl_ctx: &PossiblyCurrentContext,
+) {
+    match event {
+        Event::LoopDestroyed => (),
+        Event::WindowEvent { event, .. } => match event {
+            WindowEvent::Resized(physical_size) => {
+                // Handle window resizing
+                gl_surface.resize(
+                    gl_ctx,
+                    NonZeroU32::new(physical_size.width).unwrap(),
+                    NonZeroU32::new(physical_size.height).unwrap(),
+                );
+            }
+            WindowEvent::CloseRequested => control_flow.set_exit(),
+            WindowEvent::KeyboardInput {
+                input:
+                    KeyboardInput {
+                        virtual_keycode: Some(VirtualKeyCode::Escape),
+                        state: ElementState::Pressed,
+                        ..
+                    },
+                ..
+            } => {
+                info!("There is an Escape D:");
+                control_flow.set_exit();
+            }
+            _ => (),
+        },
+        _ => (),
+    }
 }
